pyproject.toml
[project]
name = "procrastination-police"
version = "0.1.0"
description = "A computer vision app that shame-reminds you to get back to work when you look at your phone"
dependencies = [
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]